pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel", "setuptools_scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "netnovelcrawler"
dynamic = ["version"]
license = { text = "MIT" }
authors = [{ name = "NovelReader", email = "xxxx@hotmail.com" }]
description = "Crawler framework to download Internet-novels from web."
readme = { file = "README.md", content-type = "text/markdown" }
keywords = ["net novel", "crawler"]
classifiers = [
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Entertainment :: Novel Reading",
    "Topic :: Internet :: WWW/HTTP",
    "Programming Language :: Python :: Implementation :: CPython",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: Chinese (Simplified)"]

dependencies = [
    "beautifulsoup4",
    "requests",
    "setuptools",
    "setuptools_scm",
    "selenium",
    "numpy",
    "Pillow"]

[project.optional-dependencies]
gui = ["PyQt6"]

[project.scripts]
mgrui = "netnovelcrawlertaskmgr:run_mgr"

[tool.setuptools]
packages = ["crawler", "crawlertaskmgr"]

[tool.setuptools_scm]

[tool.pylint]
disable = ["C0104", "C0114", "C0115", "C0116", "C0301", "C0302", "C0411", "C0413", "C1802", "R0401", "R0801", "R0901", "R0902", "R0903", "R0904", "R0911", "R0912", "R0913", "R0914", "R0915", "R1702", "R1720", "W0105", "W0123", "W0201", "W0511", "W0613", "W1113", "W1514", "E0401", "E1121", "C0103", "C0209", "R0402", "R1705", "R1710", "R1725", "R1735", "W0102", "W0212", "W0221", "W0223", "W0231", "W0237", "W0612", "W0621", "W0622", "W0703", "W1309", "E1102", "E1136", "I1101", "C0112", "W1201", "W1202", "W1203"]
output-format = "colorized"
extension-pkg-allow-list = ["numpy", "pandas", "plotly", "PySide6", "PyQt5", "PyQt6"]
max-line-length = 120

[tool.black]
line-length = 120
target-version = ['py39', 'py310', 'py311', 'py312']

[tool.isort]
profile = "black"
line_length = 120
known_third_party = ["numpy", "pandas", "plotly", "PySide6", "PyQt5", "PyQt6"]